fix: :recycle: refactor again hehe
