+++ components/layout/Footer.tsx
@@ -1,5 +1,5 @@
 import { SocialIcon } from "@/assets/Icons";
-import { siteConfig } from "@/configs/sites";
+import { siteConfig } from "@/config/site";
 import { Text } from "@/components/typography";
 import Link from "next/link";
