+++ config/professions.ts
